+++ assets/javascripts/kitten/components/atoms/typography/common/title-modifier-styles.js
@@ -63,18 +63,34 @@ export const titleModifiers = [
   },
 ]
 
+const getLetterSpacingFromStep = step => {
+  if (step > 5) return '-0.02em'
+  if (step > 3) return '-0.015em'
+  return '-0.01em'
+}
+const getLineHeightFromStep = step => {
+  if (step > 8) return '1em'
+  if (step > 4) return `calc(1em + ${pxToRem(4)})`
+  return `calc(1em + ${pxToRem(2)})`
+}
+
+const getStyleFromStep = step => css`
+  font-size: ${stepToRem(step)};
+  letter-spacing: ${getLetterSpacingFromStep(step)};
+  line-height: ${getLineHeightFromStep(step)};
+`
+
 export const titleModifierStyles = prefix => css`
   ${titleModifiers.map((key, index) => {
     return css`
       ${`${prefix}--${titleModifiersNames[index]}`} {
-        font-size: ${stepToRem(key.fontStepOnMobile)};
-        line-height: 1.2;
+        ${getStyleFromStep(key.fontStepOnMobile)};
 
         @media (min-width: ${pxToRem(ScreenConfig.S.min)}) {
-          font-size: ${stepToRem(key.fontStepOnTablet)};
+          ${getStyleFromStep(key.fontStepOnTablet)};
         }
         @media (min-width: ${pxToRem(ScreenConfig.L.min)}) {
-          font-size: ${stepToRem(key.fontStepOnDesktop)};
+          ${getStyleFromStep(key.fontStepOnDesktop)};
         }
       }
     `
@@ -88,15 +104,14 @@ export const titleHelperModifierStyles = modifier => {
   const modifierDefinitions = titleModifiersNames[modifierIndex]
 
   return css`
-    font-size: ${stepToRem(modifierDefinitions.fontStepOnMobile)};
-    line-height: 1.2;
+    ${getStyleFromStep(modifierDefinitions.fontStepOnMobile)};
 
     @media (min-width: ${pxToRem(ScreenConfig.S.min)}) {
-      font-size: ${stepToRem(modifierDefinitions.fontStepOnTablet)};
+      ${getStyleFromStep(modifierDefinitions.fontStepOnTablet)};
     }
 
     @media (min-width: ${pxToRem(ScreenConfig.L.min)}) {
-      font-size: ${stepToRem(modifierDefinitions.fontStepOnDesktop)};
+      ${getStyleFromStep(modifierDefinitions.fontStepOnDesktop)};
     }
   `
 }